requirements.txt
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
nltk==3.8.1
oauth2==1.9.0.post1
oauthlib==3.2.2
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4